final fix

+++ migrations/000001_create_wallets_table.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE wallets (
     id UUID PRIMARY KEY,
-    balance INTEGER NOT NULL DEFAULT 0,
+    balance DECIMAL NOT NULL DEFAULT 0,
     CONSTRAINT non_negative_balance CHECK (balance >= 0)
 );
